requirements.txt
numpy~=1.16.6
future==0.16.0
Pillow~=9.2.0
keyboard~=0.13.5
snakeviz==2.1.0
